# week4/2.Web Api using .Net core with Swagger/CODE/WebApiwithSwagger/Controllers/EmpController.cs
using Microsoft.AspNetCore.Mvc;

namespace WebApiwithSwagger.Controllers
{
    [ApiController]
    [Route("api/[controller]")]
    public class EmpController : ControllerBase
    {
        private static List<string> values = new List<string>
        {
            "Name: SAUMY RAJ",
            "Superset ID: 6363348",
            "College: KIIT"
        };

        [HttpGet]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
        public ActionResult<IEnumerable<string>> Get()
        {
            return values;
        }

        [HttpGet("{id}")]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status404NotFound)]
        public ActionResult<string> Get(int id)
        {
            if (id < 0 || id >= values.Count)
                return NotFound("ID not found");
            return values[id];
        }

        [HttpPost]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status400BadRequest)]
        public IActionResult Post([FromBody] string value)
        {
            values.Add(value);
            return Ok($"Added: {value}");
        }

        [HttpPut("{id}")]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status404NotFound)]
        public IActionResult Put(int id, [FromBody] string value)
        {
            if (id < 0 || id >= values.Count)
                return NotFound("ID not found");

            values[id] = value;
            return Ok($"Updated id {id} with value {value}");
        }

        [HttpDelete("{id}")]
        [ProducesResponseType(StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status404NotFound)]
        public IActionResult Delete(int id)
        {
            if (id < 0 || id >= values.Count)
                return NotFound("ID not found");

            values.RemoveAt(id);
            return Ok($"Deleted id {id}");
        }
    }
}